- Use multithreading for clean phase using --thread parameter (#108) - Refactored AbstractSynchronize (removed inner classes) - Fixed ident
